#79 Move localstorage to own file, also create a sessionstorage function
