+++ src/stateManagement/choosenApplication.ts
@@ -1,5 +1,5 @@
 import { atom } from "recoil";
-import { localStorageEffect } from "./userAuth";
+import { localStorageEffect } from "./localstorageRecoil";
 
 export const choosenApplicationState = atom<string>({
   key: "applicationState",
