+++ src/stateManagement/localstorageRecoil.ts
@@ -1,0 +1,18 @@
+import { DefaultValue } from "recoil";
+
+export const localStorageEffect = (key: string) => ({
+  setSelf,
+  onSet,
+}: any) => {
+  const savedValue = localStorage.getItem(key);
+  if (savedValue != null) {
+    setSelf(JSON.parse(savedValue));
+  }
+  onSet((newValue: any) => {
+    if (newValue instanceof DefaultValue) {
+      localStorage.removeItem(key);
+    } else {
+      localStorage.setItem(key, JSON.stringify(newValue));
+    }
+  });
+};
