+++ src/stateManagement/userAuth.ts
@@ -1,27 +1,9 @@
-import { atom, DefaultValue } from "recoil";
+import { atom } from "recoil";
 import firebase from "firebase/app";
-
-export const localStorageEffect = (key: string) => ({
-  setSelf,
-  onSet,
-}: any) => {
-  const savedValue = localStorage.getItem(key);
-  if (savedValue != null) {
-    setSelf(JSON.parse(savedValue));
-  }
-  onSet((newValue: any) => {
-    if (newValue instanceof DefaultValue) {
-      localStorage.removeItem(key);
-    } else {
-      localStorage.setItem(key, JSON.stringify(newValue));
-    }
-  });
-};
+import { sessionStorageEffect } from "./sessionstorageRecoil";
 
 export const currentUserState = atom<firebase.User | null>({
   key: "currentUserState",
   default: null,
-  effects_UNSTABLE: [localStorageEffect("current_user")],
+  effects_UNSTABLE: [sessionStorageEffect("current_user")],
 });
-
-
